+++ Providers/Hangfire/Revo.Hangfire/HangfireJobScheduler.cs
@@ -22,6 +22,19 @@ namespace Revo.Hangfire
             return (Task<string>)genericMethod.Invoke(this, new object[] { job, enqueueAt });
         }
 
+        public Task AddOrUpdateRecurringJobAsync(IJob job, string jobId, string cronExpression)
+        {
+            var method = GetType().GetMethod(nameof(DoAddOrUpdateRecurringJobAsync), BindingFlags.Instance | BindingFlags.NonPublic);
+            var genericMethod = method.MakeGenericMethod(job.GetType());
+            return (Task)genericMethod.Invoke(this, new object[] { job, jobId, cronExpression });
+        }
+
+        public Task RemoveRecurringJobIfExists(string jobId)
+        {
+            RecurringJob.RemoveIfExists(jobId);
+            return Task.CompletedTask;
+        }
+
         private Task<string> DoEnqeueJobAsync<TJob>(TJob job, TimeSpan? timeDelay)
             where TJob : IJob
         {
@@ -42,7 +55,17 @@ namespace Revo.Hangfire
             return Task.FromResult(jobId);
         }
 
-        public Task DeleteJobScheduleAsync(string jobId)
+        private Task DoAddOrUpdateRecurringJobAsync<TJob>(TJob job, string jobId, string cronExpression)
+            where TJob : IJob
+        {
+            RecurringJob.AddOrUpdate<HangfireJobEntryPoint<TJob>>(
+                jobId,
+                entryPoint => entryPoint.ExecuteAsync(job),
+                cronExpression);
+            return Task.CompletedTask;
+        }
+
+        public Task DeleteScheduleJobAsync(string jobId)
         {
             BackgroundJob.Delete(jobId);
             return Task.FromResult(0);